+++ src/node.rs
@@ -1,11 +1,11 @@
 #[derive(Debug)]
 pub struct Node {
-    id: i32,
+    id: usize,
     weight: f32,
 }
 
 impl Node {
-    pub fn new(id: i32, weight: f32) -> Self {
+    pub fn new(id: usize, weight: f32) -> Self {
         Node { id, weight }
     }
 
@@ -13,7 +13,7 @@ impl Node {
         self.weight
     }
 
-    pub fn get_id(&self) -> i32 {
+    pub fn get_id(&self) -> usize {
         self.id
     }
 }
